🧹 refactor: break down large files, add tests and tracing spans
🎯 What:
- Refactored `src/lib.rs` into multiple submodules under `src/api/` and re-exported them.
- Refactored `src/link/link_and_linkbuilder.rs` and `src/link/rules.rs` into modular structures with private submodules.
- Added missing tests in `src/markdown/replace_include.rs` and `src/cli/mod.rs` (checking clap assertion).
- Replaced multiple `.unwrap()` calls with safe `.context(...)` error handling in several modules.
- Added `#[tracing::instrument]` to key functions for better observability.
- Assured the code is fully formatted and linted with all docs present (`-W missing_docs`).

💡 Why:
- Keeping modules small improves code readability, maintainability and limits the file scope.
- `unwrap()` calls can cause runtime panics on edge cases; structured error handling offers graceful recovery.
- More comprehensive test coverage ensures stability for CI/CD pipelines.

✅ Verification:
- All new and existing tests successfully pass via `cargo test`.
- Ran `cargo check`, `cargo clippy -- -D warnings -W missing_docs` to ensure code is sound.

✨ Result:
- Codebase is significantly cleaner and better structured, safer against unexpected runtime states, and fully covered test-wise.
