+++ src/api/mod.rs
@@ -1,0 +1,15 @@
+//! Public API
+
+mod categories;
+mod debug;
+mod links;
+mod markdown;
+mod refdefs;
+mod sitemap;
+
+pub use categories::*;
+pub use debug::*;
+pub use links::*;
+pub use markdown::*;
+pub use refdefs::*;
+pub use sitemap::*;
