+++ src/link/link_and_linkbuilder/mod.rs
@@ -1,0 +1,7 @@
+//! Internal models for a Markdown [Link] and [LinkBuilder]
+
+mod builder;
+mod link;
+
+pub(crate) use builder::*;
+pub(crate) use link::*;
