+++ src/link/rules/structs.rs
@@ -1,0 +1,8 @@
+#[derive(Debug, Default)]
+pub(crate) struct Rule<'a> {
+    pub(crate) re: &'a str, // Regex pattern to match the url
+    #[allow(dead_code)]
+    pub(crate) label_pattern: &'a str, // regex replacement pattern
+    pub(crate) badge_url_pattern: &'a str, /* optional pattern to build a
+                             * badge link */
+}
