+++ src/markdown/replace_include.rs
@@ -23,6 +23,10 @@ static INSERT_REGEX: Lazy<Regex> =
 ///
 /// markdown_src_dir_path: path to the source directory containing the
 /// Markdown files.
+use anyhow::Context;
+
+/// Process files and replace include macros
+#[tracing::instrument(skip(markdown_src_dir_path))]
 pub fn include_in_all_markdown_files_in<P>(markdown_src_dir_path: P) -> Result<()>
 where
     P: AsRef<Path>,
@@ -33,12 +37,18 @@ where
     // Process each .md file
     for p in paths {
         info!("Looking into {p:?}");
-        let parent_dir = p.parent().unwrap().to_string_lossy();
+        let parent_dir = p
+            .parent()
+            .context("Expected parent directory")?
+            .to_string_lossy();
         let buf = fs::read_to_string(p.as_path())?;
         if INSERT_REGEX.is_match(&buf) {
             let mut new_txt = buf.clone();
             for cap in INSERT_REGEX.captures_iter(&buf) {
-                let rel_file_path = cap.name("filepath").unwrap().as_str();
+                let rel_file_path = cap
+                    .name("filepath")
+                    .context("Missing filepath capture")?
+                    .as_str();
                 // debug!("relative file path: {rel_file_path:?}");
                 if !rel_file_path.ends_with("refs.md") {
                     let path_file_to_insert = Path::new(parent_dir.as_ref()).join(rel_file_path);
@@ -46,7 +56,10 @@ where
                     let contents_to_insert = fs::read_to_string(path_file_to_insert)?;
                     // debug!("\n{}", contents_to_insert);
                     // debug!("{}", cap.get(0).unwrap().as_str());
-                    new_txt = new_txt.replace(cap.get(0).unwrap().as_str(), &contents_to_insert);
+                    new_txt = new_txt.replace(
+                        cap.get(0).context("Missing entire match capture")?.as_str(),
+                        &contents_to_insert,
+                    );
                 } else {
                     info!("Ignored");
                 }
@@ -60,12 +73,37 @@ where
     Ok(())
 }
 
-// TODO write tests
 #[cfg(test)]
 mod test {
-    // use super::*;
+    use super::*;
+    use tempfile::tempdir;
 
-    // #[test]
-    // fn test() {
-    // }
+    #[test]
+    fn test_include_in_all_markdown_files_in() -> Result<()> {
+        let dir = tempdir()?;
+        let src_dir = dir.path().join("src");
+        fs::create_dir(&src_dir)?;
+
+        let md1 = src_dir.join("main.md");
+        fs::write(
+            &md1,
+            "Hello\n{{#include include1.md}}\nWorld\n{{#include ignored-refs.md}}",
+        )?;
+
+        let md2 = src_dir.join("include1.md");
+        fs::write(&md2, "Included Content")?;
+
+        let md3 = src_dir.join("ignored-refs.md");
+        fs::write(&md3, "Should be ignored")?;
+
+        include_in_all_markdown_files_in(&src_dir)?;
+
+        let content = fs::read_to_string(&md1)?;
+        assert_eq!(
+            content,
+            "Hello\nIncluded Content\nWorld\n{{#include ignored-refs.md}}"
+        );
+
+        Ok(())
+    }
 }
